feat : add biome setting
