+++ amplify/auth/resource.ts
@@ -1,11 +1,11 @@
-import { defineAuth } from '@aws-amplify/backend';
+import { defineAuth } from "@aws-amplify/backend";
 
 /**
  * Define and configure your auth resource
  * @see https://docs.amplify.aws/gen2/build-a-backend/auth
  */
 export const auth = defineAuth({
-  loginWith: {
-    email: true,
-  },
+	loginWith: {
+		email: true,
+	},
 });
