--- app/client.ts
@@ -1,6 +1,0 @@
-import { Schema } from '@/amplify/data/resource';
-import { createAIHooks } from '@aws-amplify/ui-react-ai';
-import { generateClient } from 'aws-amplify/api';
-
-export const client = generateClient<Schema>({ authMode: 'userPool' });
-export const { useAIConversation, useAIGeneration } = createAIHooks(client);
