--- app/layout.tsx
@@ -1,25 +1,0 @@
-import { ConfigureAmplifyClientSide } from '@/components/ConfigureAmplify';
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'Personal RAG Chat',
-  description: 'Personal RAG Chat System',
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang='en'>
-      <body className={inter.className}>
-        <ConfigureAmplifyClientSide />
-        {children}
-      </body>
-    </html>
-  );
-}
