+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "format": "npx @biomejs/biome format --write src amplify/**/*.ts",
+    "lint": "npx @biomejs/biome format src amplify/**/*.ts",
     "tsc": "npx tsc --noEmit"
   },
   "dependencies": {
@@ -25,6 +26,7 @@
     "@aws-amplify/backend-cli": "1.4.6",
     "@aws-sdk/client-dynamodb": "3.726.1",
     "@aws-sdk/lib-dynamodb": "3.726.1",
+    "@biomejs/biome": "1.9.4",
     "@langchain/aws": "0.1.3",
     "@langchain/community": "0.3.24",
     "@langchain/core": "0.3.29",
@@ -35,8 +37,6 @@
     "aws-cdk-lib": "2.175.1",
     "constructs": "10.4.2",
     "esbuild": "0.24.2",
-    "eslint": "8",
-    "eslint-config-next": "14.2.23",
     "hnswlib-node": "3.0.0",
     "langchain": "0.3.11",
     "pdf-parse": "1.1.1",
