+++ renovate.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["config:base"],
+  "extends": ["config:recommended", "regexManagers:biomeVersions"],
   "packageRules": [
     {
       "groupName": "all non-major dependencies",
